Move executable jar tests to separate class
